chore: revert moving to edge
This reverts commit 9e354b9cb476bafd7938dbd1470c3f409c1624cf.

Said commit was merged for CI/CD to produce a release with a patched
version of chromium for CVE-2025-5959, which at the time of writing is
only available on alpine edge.

The release has been produced and we should go back to stable, as
staying on edge will prevent renovate from creating chromium updates.